extract FileSystemHelperFunctions out of NodeExecutionFunctions
