+++ packages/core/src/execution-engine/node-execution-context/supply-data-context.ts
@@ -23,7 +23,6 @@ import { createDeferredPromise } from 'n8n-workflow';
 import {
 	constructExecutionMetaData,
 	copyInputItems,
-	getFileSystemHelperFunctions,
 	getRequestHelperFunctions,
 	getSSHTunnelFunctions,
 	normalizeItems,
@@ -38,6 +37,7 @@ import {
 	getBinaryHelperFunctions,
 } from './utils/binary-helper-functions';
 import { getDeduplicationHelperFunctions } from './utils/deduplication-helper-functions';
+import { getFileSystemHelperFunctions } from './utils/file-system-helper-functions';
 import { getInputConnectionData } from './utils/get-input-connection-data';
 
 export class SupplyDataContext extends BaseExecuteContext implements ISupplyDataFunctions {
