+++ packages/core/src/execution-engine/node-execution-context/utils/__tests__/file-system-helper-functions.test.ts
@@ -1,0 +1,13 @@
+import { Container } from '@n8n/di';
+
+import { InstanceSettings } from '@/instance-settings';
+
+import { isFilePathBlocked } from '../file-system-helper-functions';
+
+describe('isFilePathBlocked', () => {
+	test('should return true for static cache dir', () => {
+		const filePath = Container.get(InstanceSettings).staticCacheDir;
+
+		expect(isFilePathBlocked(filePath)).toBe(true);
+	});
+});
